backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
docker
python-whois
Jinja2